new channel data model [no ci]
Our current model for channel data is flawed because `ChannelState`
contains data that is unrelated to the actual channel state. This is
why deserializing channel data requires an `NodeParams` to be provided,
instead of simply being a piece of data that can be serialized and
deserialized without context.

The non-channel-related data in question already stands out as the only
`abstract` values at the top of the `ChannelState` hierarchy, and they
are the same for all channel states. In eclair, those would typically be
provided in the constructor of the `FSM` actor (as opposed to attributes
of the data class).

In this commit, we separate the channel state from the context in which
the channel operates (which we could also call the Machine of the FSM).
For clarity, below is the gist of the change. The `State`/`Command`/
`Action` triplet remains the same, but we have a new `Context`, which
carries external context (node params, current block count, etc.).
Because `process` is now an _extension_ function of `Context`, it has
access to both the context and state data from its body. This is what
makes the change almost a drop-in replacement. Running a command on a
state is now slightly more complex, but it is also closer to the eclair
codebase.

```scala
sealed class Command {
    object Command1 : Command()
}

sealed class Action {
    object Action1 : Action()
}

sealed class State {

    abstract fun Context.process(cmd: Command): Pair<State, List<Action>>

    data class State1(val a: String) : State() {
        override fun Context.process(cmd: Command): Pair<State, List<Action>> {
            // we have access to both the context and the state
            val params = this.params
            val state = this@State1
            return State2(10) to listOf<Action.Action1>()
        }
    }

    data class State2(val b: Int) : State() {
        override fun Context.process(cmd: Command): Pair<State, List<Action>> {
            TODO("Not yet implemented")
        }
    }
}

data class Params(val alias: String)

data class Context(val params: Params)

fun test() {
    val context = Context(Params(alias = "my node"))
    val state = State.State1("foobar")

    val (state1, actions) = state.run { context.process(Command.Command1) }
}
```

Notes:
- overall there is less code than before because we remove unnecessary
data that we were carrying around for no reason
- end states like `Aborted` are now `object`s
- the channel state containing `currentTip` was misleading, because it
was not consistent with pending htlcs