Rework and simplify electrum client, miniwallet and watcher
CompletableDeferred<> is used to handle multiple concurrent calls and provide suspend methods which are much easier to use.
For example, to retrieve a transaction you can simply call `client.getTx(txid)` and you can chain such calls in your message handling loops:
this eliminate the need to send messages to yourself and create intermediate states that are just waiting for a response from the electrum client.